# crypto/sqlstore: add index to make finding megolm sessions to backup faster