--- a/crypto/sql_store_upgrade/18-megolm-inbound-session-backup-index.sql
+++ b/crypto/sql_store_upgrade/18-megolm-inbound-session-backup-index.sql
@@ -1,2 +1,2 @@
 -- v18 (compatible with v15+): Add an index to the megolm_inbound_session table to make finding sessions to backup faster
-CREATE INDEX crypto_megolm_inbound_session_backup_idx ON crypto_megolm_inbound_session(account_id, key_backup_version, session);
+CREATE INDEX crypto_megolm_inbound_session_backup_idx ON crypto_megolm_inbound_session(account_id, key_backup_version) WHERE session IS NOT NULL;